+++ TwitchChannelPointsMiner/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-__version__ = "3.3.0"
+__version__ = "3.4.0"
 from .TwitchChannelPointsMiner import TwitchChannelPointsMiner
 
 __all__ = [